feat: (wip) update chart styles
